Moved net plugin to its own directory
